third_party/WebKit/LayoutTests/inspector-enabled/sources/debugger/script-formatter-console-expected.txt

CONSOLE ERROR: line 12: Test message.
Tests that the script formatting changes console line numbers.


Running: testSetup

Running: testConsoleMessagesForFormattedScripts
script-formatter-console.html:12 
Pre-format row message list:
[11]
script-formatter-console.html:formatted:15
Post-format row message list:
[14]
